+++ src/redux/reducers/appReducer.ts
@@ -1,7 +1,6 @@
 import {loginMeThunk} from "./authReducer";
-import {Dispatch} from 'redux';
 import {ThunkDispatch} from "redux-thunk";
-import {AppStateType} from "../redux-store";
+import {AppStateType, InferActionsTypes} from "../redux-store";
 
 // Name action
 const SET_INITIALIZED = 'SET_INITIALIZED'
@@ -27,21 +26,22 @@ const appReducer = (state = initialState, action: ActionsType): InitialStateInte
     }
 }
 
-interface InitializingACInterface {
-    type: typeof SET_INITIALIZED
-}
-
-type ActionsType = InitializingACInterface
+type ActionsType = InferActionsTypes<typeof appAction>
 type DispatchThunkType = ThunkDispatch<AppStateType, unknown, ActionsType>
 
+
 // actionCreator
-const initializingAC = ():InitializingACInterface => ({type:SET_INITIALIZED})
+export const appAction = {
+    initializingAC:() => ({type:SET_INITIALIZED} as const)
+}
+
 
 // thunkCreator
 export const initializeApp = () => async (dispatch: DispatchThunkType) => {
     let promise = dispatch(loginMeThunk())
     await Promise.all([promise])
-    dispatch(initializingAC())
+    dispatch(appAction.initializingAC())
 }
+
 
 export default appReducer
